+++ bindings/angular2/examples/range.spec.js
@@ -21,4 +21,17 @@ describe('range.html', () => {
 
   //   expect($('#value').getText()).not.toBe('10');
   // });
+
+  describe('reative form', () => {
+    it('should change form value when the range value is changed', () => {
+      browser.actions()
+        .mouseMove($('#reactive-range'))
+        .mouseDown()
+        .mouseMove({x: 500, y: 0}) // drag the slider a long way right and hope we reached the edge
+        .mouseUp()
+        .perform();
+
+      expect($('#reactive-value').getText()).toBe('100');
+    });
+  });
 });