+++ CHANGELOG.md
@@ -8,6 +8,7 @@ dev
  ### New Features
 
  * ons.notification: Add `maskColor` configuration for `alert`, `confirm` and `prompt` ([#2358](https://github.com/OnsenUI/OnsenUI/issues/2358))
+ * angular2: Add support for reactive forms (PR [#2401](https://github.com/OnsenUI/OnsenUI/pull/2401). Fixes [#2261](https://github.com/OnsenUI/OnsenUI/issues/2261).)
 
  ### Bug Fixes